requirements.txt
django
django-allauth
pytest
pytest-django
pytest-cov
pytz
